+++ docs/releases.md
@@ -14,6 +14,7 @@
 * Option to plot bathymetry along fixed lat/lon section ([#399](https://github.com/CWorthy-ocean/roms-tools/pull/399))
 * Truly centered Gaussian CDR releases, reflecting recent changes in ucla-roms ([#413](https://github.com/CWorthy-ocean/roms-tools/pull/413))
 * Add `compute_total_cdr_source ` method to `CDRForcing`, which computes integrated tracer quantities for all releases and returns a pandas DataFrame including a units row ([#444](https://github.com/CWorthy-ocean/roms-tools/pull/444))
+* Add utility function `get_glorys_bounds` to return the min/max latitude and longitude required for GLORYS downloads for a fixed domain ([#445](https://github.com/CWorthy-ocean/roms-tools/pull/445))
 
 ### Breaking Changes
 
@@ -22,6 +23,8 @@
 ### Internal Changes
 
 ### Documentation
+
+* Add an end-to-end example to help users run through the full workflow ([#468](https://github.com/CWorthy-ocean/roms-tools/pull/468))
 
 ### Bugfixes
 
@@ -43,6 +46,7 @@
 
 * Fix broken link to GLORYS dataset ([#397](https://github.com/CWorthy-ocean/roms-tools/pull/397))
 * Include recommendation about `interp_frc` in INFO message ([#352](https://github.com/CWorthy-ocean/roms-tools/issues/352))
+* Document downloading of external datasets ([#445](https://github.com/CWorthy-ocean/roms-tools/pull/445))
 
 ## v3.1.1